fmt + snapshot + fuzz test

+++ test/mocks/uniswap/MinimalObservablePool.sol
@@ -8,15 +8,10 @@ contract MinimalObservablePool {
         tickCumulatives = _tickCumulatives;
     }
 
-
-    function observe(uint32[] calldata secondsAgos)
-        external
-        view
-        returns (int56[] memory, uint160[] memory)
-    {
+    function observe(uint32[] calldata secondsAgos) external view returns (int56[] memory, uint160[] memory) {
         int56[] memory _tickCumulatives = new int56[](secondsAgos.length);
         uint160[] memory _secondsPerLiquidityCumulativeX128s = new uint160[](secondsAgos.length);
-        for (uint i = 0; i < secondsAgos.length; i++) {
+        for (uint256 i = 0; i < secondsAgos.length; i++) {
             _tickCumulatives[i] = tickCumulatives[i];
             _secondsPerLiquidityCumulativeX128s[i] = 0;
         }
